+++ src/rinoh/font/__init__.py
@@ -198,13 +198,13 @@ class Typeface(Resource, dict):
             Font: the next font in this typeface
 
         """
-        for width in (w for w in FontWidth if w in self):
-            for slant in (s for s in FontSlant if s in self[width]):
-                for weight in (w for w in FontWeight
-                               if w in self[width][slant]):
+        for width in sorted(self):
+            for slant in self[width]:
+                for weight in sorted(self[width][slant]):
                     yield self[width][slant][weight]
 
-    def get_font(self, weight='medium', slant='upright', width='normal'):
+    def get_font(self, weight=FontWeight.REGULAR, slant=FontSlant.UPRIGHT,
+                 width=FontWidth.NORMAL):
         """Return the font matching or closest to the given style
 
         If a font with the given weight, slant and width is available, return
@@ -219,37 +219,22 @@ class Typeface(Resource, dict):
             Font: the requested font
 
         """
-        def find_closest_style(style, styles, alternatives):
-            try:
-                return style, styles[style]
-            except KeyError:
-                for option in alternatives[style]:
-                    try:
-                        return option, styles[option]
-                    except KeyError:
-                        continue
+        def find_closest(attribute, value, available):
+            nearest = attribute.nearest(value, available)
+            return nearest, available[nearest]
 
-        def find_closest_weight(weight, weights):
-            index = FontWeight.values.index(weight)
-            min_distance = len(FontWeight.values)
-            closest = None
-            for i, option in enumerate(FontWeight.values):
-                if option in weights and abs(index - i) < min_distance:
-                    min_distance = abs(index - i)
-                    closest = option
-            return closest, weights[closest]
-
-        available_width, slants = find_closest_style(width, self,
-                                                     FontWidth.alternatives)
-        available_slant, weights = find_closest_style(slant, slants,
-                                                      FontSlant.alternatives)
-        available_weight, font = find_closest_weight(weight, weights)
-
-        if (available_width != width or available_slant != slant or
-            available_weight != weight):
+        available_width, slants = find_closest(FontWidth, width, self)
+        available_slant, weights = find_closest(FontSlant, slant, slants)
+        available_weight, font = find_closest(FontWeight, weight, weights)
+        if ((available_width, available_slant, available_weight)
+                != (width, slant, weight)):
+            gv_width =  FontWidth.to_name(width)
+            gv_weigth = FontWeight.to_name(weight)
+            av_width = FontWidth.to_name(available_width)
+            av_weight = FontWeight.to_name(available_weight)
             warn('{} does not include a {} {} {} font. Falling back to {} {} '
-                 '{}'.format(self.name, width, weight, slant, available_width,
-                             available_weight, available_slant))
+                 '{}'.format(self.name, gv_width, gv_weigth, slant, av_width,
+                             av_weight, available_slant))
         return font
 
     # TODO: return bolder font than given font